Multiblock implementation
