+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/base/Multiblock.kt
@@ -9,13 +9,11 @@ import org.bukkit.block.Block
  * Multiblocks are more difficult than normal Pylon blocks for the simple reason that a multiblock
  * may contain some blocks that have not been loaded because they are in a different chunk.
  *
- * Ticking multiblocks should only tick when isFormedAndFullyLoaded() returns true, to avoid ticking
- * a multiblock that is either not formed, or not fully loaded (ie, not all of its components are
- * loaded).
+ * Ticking multiblocks should only tick when isFormedAndFullyLoaded() returns true, to avoid
+ * ticking a multiblock that is either not formed, or not fully loaded (ie, not all of its components
+ * are loaded).
  */
 interface Multiblock {
-    var formed: Boolean
-
     // This is automatically implemented by PylonBlock (lol)
     val block: Block
 
@@ -25,11 +23,19 @@ interface Multiblock {
     val chunksOccupied: Set<ChunkPosition>
 
     /**
-     * Check whether the multiblock is formed, and update the formed variable accordingly.
+     * Warning: You should not call this yourself. If you want to know if the multiblock is formed,
+     * use isFormedAndFullyLoaded(), which will read from a cache, so is much faster.
      *
      * You can assume that when this method is called, all components of the multiblock are loaded.
      */
-    fun refresh()
+    fun checkFormed(): Boolean
+
+    /**
+     * Returns true if every component of the multiblock is loaded AND the multiblock is formed.
+     * Reads from a cache, so call as often as you want.
+     */
+    fun isFormedAndFullyLoaded(): Boolean
+            = MultiblockCache.isFormed(this)
 
     /**
      * Should return true if there is any scenario in which this block could be part of a formed
@@ -38,28 +44,4 @@ interface Multiblock {
      * This could be called often, so make it lightweight.
      */
     fun isPartOfMultiblock(otherBlock: Block): Boolean
-
-    /**
-     * Convenience method. You should use this to check if the multiblock can be ticked.
-     */
-    fun isFormedAndFullyLoaded(): Boolean
-        = formed && isFullyLoaded()
-
-    /**
-     * Returns true if all the chunks that this multiblock can occupy are loaded.
-     */
-    fun isFullyLoaded(): Boolean
-        = MultiblockCache.isFullyLoaded(this)
-
-    /**
-     * Called when a block in any of the multiblock's chunks is modified (eg: water turning into ice).
-     *
-     * You can assume that when this method is called, all components of the multiblock are loaded.
-     *
-     * You cannot assume that the change has been made yet. For example, if ice melts into water, the
-     * supplied otherBlock will be of type water, but the block in-world may still be of type ice.
-     *
-     * You can that any BlockStorage updates have already taken place.
-     */
-    fun onComponentModified(newBlock: Block)
 }
